+++ CHANGES.rst
@@ -15,7 +15,13 @@ New:
 
 Fixes:
 
-- *add item here*
+- fix https://github.com/plone/Products.CMFPlone/issues/2469:
+  "Subobjects are indexing attributes of parent".
+  Allow only direct attributes and acquired PythonScripts,
+  but not aqcquired attributes.
+  Indexers and PythonScripts are able to handle this explicitly,
+  because they get the acquisition-wrapped object.
+  [jensens]
 
 
 1.0.4 (2016-02-25)